Library added

+++ src/Member.cpp
@@ -8,7 +8,7 @@ std::vector<Book> Member::checkBooks() const { return this->borrowedBooks; }
 void Member::borrowBook(Book book) { this->borrowedBooks.push_back(book); }
 
 void Member::returnBook(Book book) {
-  for (int i = 0; i < this->borrowedBooks.size(); i++) {
+  for (int i = 0; i < (int)this->borrowedBooks.size(); i++) {
     if (this->borrowedBooks[i] == book) {
       this->borrowedBooks.erase(this->borrowedBooks.begin() + i);
       break;
